Refactorings, placing virus scanner and fileupload in separate directories
